extract ttransaction-status-client-types and transform associated methods to free funcs: UiInstruction::parse -> parse_ui_instruction; UiInnerInstructions::parse => parse_ui_inner_instructions; UiTransactionStatusMeta::parse -> parse_ui_transaction_status_meta
