+++ ledger-tool/src/main.rs
@@ -75,7 +75,7 @@ use {
         transaction::{MessageHash, SanitizedTransaction, SimpleAddressLoader},
     },
     solana_stake_program::{points::PointValue, stake_state},
-    solana_transaction_status::UiInstruction,
+    solana_transaction_status::parse_ui_instruction,
     solana_unified_scheduler_pool::DefaultSchedulerPool,
     solana_vote_program::{
         self,
@@ -694,7 +694,7 @@ fn record_transactions(
                     let instructions = message
                         .instructions()
                         .iter()
-                        .map(|ix| UiInstruction::parse(ix, &message.account_keys(), None))
+                        .map(|ix| parse_ui_instruction(ix, &message.account_keys(), None))
                         .collect();
 
                     let is_simple_vote_tx = tx.is_simple_vote_transaction();
