+++ transaction-status-client-types/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "solana-transaction-status-client-types"
+description = "Core types for solana-transaction-status"
+documentation = "https://docs.rs/solana-transaction-status-client-types"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+base64 = { workspace = true }
+bincode = { workspace = true }
+bs58 = { workspace = true }
+serde = { workspace = true }
+serde_derive = { workspace = true }
+serde_json = { workspace = true }
+solana-account-decoder = { workspace = true }
+solana-signature = { workspace = true, default-features = false }
+solana-sdk = { workspace = true }
+thiserror = { workspace = true }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
