+++ transaction-status-client-types/src/lib.rs
@@ -1,0 +1,571 @@
+//! Core types for solana-transaction-status
+use {
+    crate::option_serializer::OptionSerializer,
+    base64::{prelude::BASE64_STANDARD, Engine},
+    core::fmt,
+    serde_derive::{Deserialize, Serialize},
+    serde_json::Value,
+    solana_account_decoder::parse_token::UiTokenAmount,
+    solana_sdk::{
+        instruction::CompiledInstruction,
+        message::{
+            v0::{LoadedAddresses, MessageAddressTableLookup},
+            MessageHeader,
+        },
+        reward_type::RewardType,
+        transaction::{
+            Result as TransactionResult, TransactionError, TransactionVersion, VersionedTransaction,
+        },
+        transaction_context::TransactionReturnData,
+    },
+    solana_signature::Signature,
+    thiserror::Error,
+};
+pub mod option_serializer;
+
+#[derive(Serialize, Deserialize, Clone, Copy, Debug, Eq, Hash, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub enum TransactionBinaryEncoding {
+    Base58,
+    Base64,
+}
+
+#[derive(Serialize, Deserialize, Clone, Copy, Debug, Eq, Hash, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub enum UiTransactionEncoding {
+    Binary, // Legacy. Retained for RPC backwards compatibility
+    Base64,
+    Base58,
+    Json,
+    JsonParsed,
+}
+
+impl UiTransactionEncoding {
+    pub fn into_binary_encoding(&self) -> Option<TransactionBinaryEncoding> {
+        match self {
+            Self::Binary | Self::Base58 => Some(TransactionBinaryEncoding::Base58),
+            Self::Base64 => Some(TransactionBinaryEncoding::Base64),
+            _ => None,
+        }
+    }
+}
+
+impl fmt::Display for UiTransactionEncoding {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        let v = serde_json::to_value(self).map_err(|_| fmt::Error)?;
+        let s = v.as_str().ok_or(fmt::Error)?;
+        write!(f, "{s}")
+    }
+}
+
+#[derive(Debug, Clone, Copy, Eq, Hash, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub enum TransactionDetails {
+    Full,
+    Signatures,
+    None,
+    Accounts,
+}
+
+impl Default for TransactionDetails {
+    fn default() -> Self {
+        Self::Full
+    }
+}
+
+#[derive(Error, Debug, PartialEq, Eq, Clone)]
+pub enum EncodeError {
+    #[error("Encoding does not support transaction version {0}")]
+    UnsupportedTransactionVersion(u8),
+}
+
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct ConfirmedTransactionStatusWithSignature {
+    pub signature: Signature,
+    pub slot: u64,
+    pub err: Option<TransactionError>,
+    pub memo: Option<String>,
+    pub block_time: Option<i64>,
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub enum TransactionConfirmationStatus {
+    Processed,
+    Confirmed,
+    Finalized,
+}
+
+#[derive(Debug, PartialEq, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct UiConfirmedBlock {
+    pub previous_blockhash: String,
+    pub blockhash: String,
+    pub parent_slot: u64,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub transactions: Option<Vec<EncodedTransactionWithStatusMeta>>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub signatures: Option<Vec<String>>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub rewards: Option<Rewards>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub num_reward_partitions: Option<u64>,
+    pub block_time: Option<i64>,
+    pub block_height: Option<u64>,
+}
+
+/// A duplicate representation of a Transaction for pretty JSON serialization
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct UiTransaction {
+    pub signatures: Vec<String>,
+    pub message: UiMessage,
+}
+
+/// A duplicate representation of a Message, in parsed format, for pretty JSON serialization
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct UiParsedMessage {
+    pub account_keys: Vec<ParsedAccount>,
+    pub recent_blockhash: String,
+    pub instructions: Vec<UiInstruction>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub address_table_lookups: Option<Vec<UiAddressTableLookup>>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct ParsedAccount {
+    pub pubkey: String,
+    pub writable: bool,
+    pub signer: bool,
+    pub source: Option<ParsedAccountSource>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub enum ParsedAccountSource {
+    Transaction,
+    LookupTable,
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase", untagged)]
+pub enum UiMessage {
+    Parsed(UiParsedMessage),
+    Raw(UiRawMessage),
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase", untagged)]
+pub enum EncodedTransaction {
+    LegacyBinary(String), // Old way of expressing base-58, retained for RPC backwards compatibility
+    Binary(String, TransactionBinaryEncoding),
+    Json(UiTransaction),
+    Accounts(UiAccountsList),
+}
+
+impl EncodedTransaction {
+    pub fn decode(&self) -> Option<VersionedTransaction> {
+        let (blob, encoding) = match self {
+            Self::Json(_) | Self::Accounts(_) => return None,
+            Self::LegacyBinary(blob) => (blob, TransactionBinaryEncoding::Base58),
+            Self::Binary(blob, encoding) => (blob, *encoding),
+        };
+
+        let transaction: Option<VersionedTransaction> = match encoding {
+            TransactionBinaryEncoding::Base58 => bs58::decode(blob)
+                .into_vec()
+                .ok()
+                .and_then(|bytes| bincode::deserialize(&bytes).ok()),
+            TransactionBinaryEncoding::Base64 => BASE64_STANDARD
+                .decode(blob)
+                .ok()
+                .and_then(|bytes| bincode::deserialize(&bytes).ok()),
+        };
+
+        transaction.filter(|transaction| transaction.sanitize().is_ok())
+    }
+}
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct EncodedTransactionWithStatusMeta {
+    pub transaction: EncodedTransaction,
+    pub meta: Option<UiTransactionStatusMeta>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub version: Option<TransactionVersion>,
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Reward {
+    pub pubkey: String,
+    pub lamports: i64,
+    pub post_balance: u64, // Account balance in lamports after `lamports` was applied
+    pub reward_type: Option<RewardType>,
+    pub commission: Option<u8>, // Vote account commission when the reward was credited, only present for voting and staking rewards
+}
+
+pub type Rewards = Vec<Reward>;
+
+/// A duplicate representation of a MessageAddressTableLookup, in raw format, for pretty JSON serialization
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct UiAddressTableLookup {
+    pub account_key: String,
+    pub writable_indexes: Vec<u8>,
+    pub readonly_indexes: Vec<u8>,
+}
+
+impl From<&MessageAddressTableLookup> for UiAddressTableLookup {
+    fn from(lookup: &MessageAddressTableLookup) -> Self {
+        Self {
+            account_key: lookup.account_key.to_string(),
+            writable_indexes: lookup.writable_indexes.clone(),
+            readonly_indexes: lookup.readonly_indexes.clone(),
+        }
+    }
+}
+
+/// A duplicate representation of TransactionStatusMeta with `err` field
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct UiTransactionStatusMeta {
+    pub err: Option<TransactionError>,
+    pub status: TransactionResult<()>, // This field is deprecated.  See https://github.com/solana-labs/solana/issues/9302
+    pub fee: u64,
+    pub pre_balances: Vec<u64>,
+    pub post_balances: Vec<u64>,
+    #[serde(
+        default = "OptionSerializer::none",
+        skip_serializing_if = "OptionSerializer::should_skip"
+    )]
+    pub inner_instructions: OptionSerializer<Vec<UiInnerInstructions>>,
+    #[serde(
+        default = "OptionSerializer::none",
+        skip_serializing_if = "OptionSerializer::should_skip"
+    )]
+    pub log_messages: OptionSerializer<Vec<String>>,
+    #[serde(
+        default = "OptionSerializer::none",
+        skip_serializing_if = "OptionSerializer::should_skip"
+    )]
+    pub pre_token_balances: OptionSerializer<Vec<UiTransactionTokenBalance>>,
+    #[serde(
+        default = "OptionSerializer::none",
+        skip_serializing_if = "OptionSerializer::should_skip"
+    )]
+    pub post_token_balances: OptionSerializer<Vec<UiTransactionTokenBalance>>,
+    #[serde(
+        default = "OptionSerializer::none",
+        skip_serializing_if = "OptionSerializer::should_skip"
+    )]
+    pub rewards: OptionSerializer<Rewards>,
+    #[serde(
+        default = "OptionSerializer::skip",
+        skip_serializing_if = "OptionSerializer::should_skip"
+    )]
+    pub loaded_addresses: OptionSerializer<UiLoadedAddresses>,
+    #[serde(
+        default = "OptionSerializer::skip",
+        skip_serializing_if = "OptionSerializer::should_skip"
+    )]
+    pub return_data: OptionSerializer<UiTransactionReturnData>,
+    #[serde(
+        default = "OptionSerializer::skip",
+        skip_serializing_if = "OptionSerializer::should_skip"
+    )]
+    pub compute_units_consumed: OptionSerializer<u64>,
+}
+
+impl From<TransactionStatusMeta> for UiTransactionStatusMeta {
+    fn from(meta: TransactionStatusMeta) -> Self {
+        Self {
+            err: meta.status.clone().err(),
+            status: meta.status,
+            fee: meta.fee,
+            pre_balances: meta.pre_balances,
+            post_balances: meta.post_balances,
+            inner_instructions: meta
+                .inner_instructions
+                .map(|ixs| ixs.into_iter().map(Into::into).collect())
+                .into(),
+            log_messages: meta.log_messages.into(),
+            pre_token_balances: meta
+                .pre_token_balances
+                .map(|balance| balance.into_iter().map(Into::into).collect())
+                .into(),
+            post_token_balances: meta
+                .post_token_balances
+                .map(|balance| balance.into_iter().map(Into::into).collect())
+                .into(),
+            rewards: meta.rewards.into(),
+            loaded_addresses: Some(UiLoadedAddresses::from(&meta.loaded_addresses)).into(),
+            return_data: OptionSerializer::or_skip(
+                meta.return_data.map(|return_data| return_data.into()),
+            ),
+            compute_units_consumed: OptionSerializer::or_skip(meta.compute_units_consumed),
+        }
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct UiTransactionReturnData {
+    pub program_id: String,
+    pub data: (String, UiReturnDataEncoding),
+}
+
+impl Default for UiTransactionReturnData {
+    fn default() -> Self {
+        Self {
+            program_id: String::default(),
+            data: (String::default(), UiReturnDataEncoding::Base64),
+        }
+    }
+}
+
+impl From<TransactionReturnData> for UiTransactionReturnData {
+    fn from(return_data: TransactionReturnData) -> Self {
+        Self {
+            program_id: return_data.program_id.to_string(),
+            data: (
+                BASE64_STANDARD.encode(return_data.data),
+                UiReturnDataEncoding::Base64,
+            ),
+        }
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone, Copy, Debug, Eq, Hash, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub enum UiReturnDataEncoding {
+    Base64,
+}
+
+/// A duplicate representation of LoadedAddresses
+#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct UiLoadedAddresses {
+    pub writable: Vec<String>,
+    pub readonly: Vec<String>,
+}
+
+impl From<&LoadedAddresses> for UiLoadedAddresses {
+    fn from(loaded_addresses: &LoadedAddresses) -> Self {
+        Self {
+            writable: loaded_addresses
+                .writable
+                .iter()
+                .map(ToString::to_string)
+                .collect(),
+            readonly: loaded_addresses
+                .readonly
+                .iter()
+                .map(ToString::to_string)
+                .collect(),
+        }
+    }
+}
+
+#[derive(Clone, Debug, PartialEq)]
+pub struct TransactionTokenBalance {
+    pub account_index: u8,
+    pub mint: String,
+    pub ui_token_amount: UiTokenAmount,
+    pub owner: String,
+    pub program_id: String,
+}
+
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct UiTransactionTokenBalance {
+    pub account_index: u8,
+    pub mint: String,
+    pub ui_token_amount: UiTokenAmount,
+    #[serde(
+        default = "OptionSerializer::skip",
+        skip_serializing_if = "OptionSerializer::should_skip"
+    )]
+    pub owner: OptionSerializer<String>,
+    #[serde(
+        default = "OptionSerializer::skip",
+        skip_serializing_if = "OptionSerializer::should_skip"
+    )]
+    pub program_id: OptionSerializer<String>,
+}
+
+impl From<TransactionTokenBalance> for UiTransactionTokenBalance {
+    fn from(token_balance: TransactionTokenBalance) -> Self {
+        Self {
+            account_index: token_balance.account_index,
+            mint: token_balance.mint,
+            ui_token_amount: token_balance.ui_token_amount,
+            owner: if !token_balance.owner.is_empty() {
+                OptionSerializer::Some(token_balance.owner)
+            } else {
+                OptionSerializer::Skip
+            },
+            program_id: if !token_balance.program_id.is_empty() {
+                OptionSerializer::Some(token_balance.program_id)
+            } else {
+                OptionSerializer::Skip
+            },
+        }
+    }
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct UiAccountsList {
+    pub signatures: Vec<String>,
+    pub account_keys: Vec<ParsedAccount>,
+}
+
+/// A duplicate representation of a Message, in raw format, for pretty JSON serialization
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct UiRawMessage {
+    pub header: MessageHeader,
+    pub account_keys: Vec<String>,
+    pub recent_blockhash: String,
+    pub instructions: Vec<UiCompiledInstruction>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub address_table_lookups: Option<Vec<UiAddressTableLookup>>,
+}
+
+/// A duplicate representation of a CompiledInstruction for pretty JSON serialization
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct UiCompiledInstruction {
+    pub program_id_index: u8,
+    pub accounts: Vec<u8>,
+    pub data: String,
+    pub stack_height: Option<u32>,
+}
+
+impl UiCompiledInstruction {
+    pub fn from(instruction: &CompiledInstruction, stack_height: Option<u32>) -> Self {
+        Self {
+            program_id_index: instruction.program_id_index,
+            accounts: instruction.accounts.clone(),
+            data: bs58::encode(&instruction.data).into_string(),
+            stack_height,
+        }
+    }
+}
+
+/// A duplicate representation of an Instruction for pretty JSON serialization
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase", untagged)]
+pub enum UiInstruction {
+    Compiled(UiCompiledInstruction),
+    Parsed(UiParsedInstruction),
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase", untagged)]
+pub enum UiParsedInstruction {
+    Parsed(ParsedInstruction),
+    PartiallyDecoded(UiPartiallyDecodedInstruction),
+}
+
+/// A partially decoded CompiledInstruction that includes explicit account addresses
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct UiPartiallyDecodedInstruction {
+    pub program_id: String,
+    pub accounts: Vec<String>,
+    pub data: String,
+    pub stack_height: Option<u32>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(rename_all = "camelCase")]
+pub struct ParsedInstruction {
+    pub program: String,
+    pub program_id: String,
+    pub parsed: Value,
+    pub stack_height: Option<u32>,
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct UiInnerInstructions {
+    /// Transaction instruction index
+    pub index: u8,
+    /// List of inner instructions
+    pub instructions: Vec<UiInstruction>,
+}
+
+impl From<InnerInstructions> for UiInnerInstructions {
+    fn from(inner_instructions: InnerInstructions) -> Self {
+        Self {
+            index: inner_instructions.index,
+            instructions: inner_instructions
+                .instructions
+                .iter()
+                .map(
+                    |InnerInstruction {
+                         instruction: ix,
+                         stack_height,
+                     }| {
+                        UiInstruction::Compiled(UiCompiledInstruction::from(ix, *stack_height))
+                    },
+                )
+                .collect(),
+        }
+    }
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+pub struct InnerInstructions {
+    /// Transaction instruction index
+    pub index: u8,
+    /// List of inner instructions
+    pub instructions: Vec<InnerInstruction>,
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+pub struct InnerInstruction {
+    /// Compiled instruction
+    pub instruction: CompiledInstruction,
+    /// Invocation stack height of the instruction,
+    pub stack_height: Option<u32>,
+}
+
+#[derive(Clone, Debug, PartialEq)]
+pub struct TransactionStatusMeta {
+    pub status: TransactionResult<()>,
+    pub fee: u64,
+    pub pre_balances: Vec<u64>,
+    pub post_balances: Vec<u64>,
+    pub inner_instructions: Option<Vec<InnerInstructions>>,
+    pub log_messages: Option<Vec<String>>,
+    pub pre_token_balances: Option<Vec<TransactionTokenBalance>>,
+    pub post_token_balances: Option<Vec<TransactionTokenBalance>>,
+    pub rewards: Option<Rewards>,
+    pub loaded_addresses: LoadedAddresses,
+    pub return_data: Option<TransactionReturnData>,
+    pub compute_units_consumed: Option<u64>,
+}
+
+impl Default for TransactionStatusMeta {
+    fn default() -> Self {
+        Self {
+            status: Ok(()),
+            fee: 0,
+            pre_balances: vec![],
+            post_balances: vec![],
+            inner_instructions: None,
+            log_messages: None,
+            pre_token_balances: None,
+            post_token_balances: None,
+            rewards: None,
+            loaded_addresses: LoadedAddresses::default(),
+            return_data: None,
+            compute_units_consumed: None,
+        }
+    }
+}
