+++ transaction-status/src/lib.rs
@@ -1,22 +1,35 @@
 #![allow(clippy::arithmetic_side_effects)]
 
-pub use {crate::extract_memos::extract_and_fmt_memos, solana_sdk::reward_type::RewardType};
+pub use {
+    crate::extract_memos::extract_and_fmt_memos,
+    solana_sdk::reward_type::RewardType,
+    solana_transaction_status_client_types::{
+        option_serializer, ConfirmedTransactionStatusWithSignature, EncodeError,
+        EncodedTransaction, EncodedTransactionWithStatusMeta, InnerInstruction, InnerInstructions,
+        Reward, Rewards, TransactionBinaryEncoding, TransactionConfirmationStatus,
+        TransactionDetails, TransactionStatusMeta, TransactionTokenBalance, UiAccountsList,
+        UiAddressTableLookup, UiCompiledInstruction, UiConfirmedBlock, UiInnerInstructions,
+        UiInstruction, UiLoadedAddresses, UiMessage, UiParsedInstruction, UiParsedMessage,
+        UiPartiallyDecodedInstruction, UiRawMessage, UiReturnDataEncoding, UiTransaction,
+        UiTransactionEncoding, UiTransactionReturnData, UiTransactionStatusMeta,
+        UiTransactionTokenBalance,
+    },
+};
 use {
     crate::{
         option_serializer::OptionSerializer,
-        parse_accounts::{parse_legacy_message_accounts, parse_v0_message_accounts, ParsedAccount},
-        parse_instruction::{parse, ParsedInstruction},
+        parse_accounts::{parse_legacy_message_accounts, parse_v0_message_accounts},
+        parse_instruction::parse,
     },
     base64::{prelude::BASE64_STANDARD, Engine},
-    solana_account_decoder::parse_token::UiTokenAmount,
     solana_sdk::{
         clock::{Slot, UnixTimestamp},
         commitment_config::CommitmentConfig,
         hash::Hash,
         instruction::CompiledInstruction,
         message::{
-            v0::{self, LoadedAddresses, LoadedMessage, MessageAddressTableLookup},
-            AccountKeys, Message, MessageHeader, VersionedMessage,
+            v0::{self, LoadedAddresses, LoadedMessage},
+            AccountKeys, Message, VersionedMessage,
         },
         pubkey::Pubkey,
         reserved_account_keys::ReservedAccountKeys,
@@ -25,9 +38,8 @@ use {
             Result as TransactionResult, Transaction, TransactionError, TransactionVersion,
             VersionedTransaction,
         },
-        transaction_context::TransactionReturnData,
     },
-    std::{collections::HashSet, fmt},
+    std::collections::HashSet,
     thiserror::Error,
 };
 
@@ -37,7 +49,6 @@ extern crate lazy_static;
 extern crate serde_derive;
 
 pub mod extract_memos;
-pub mod option_serializer;
 pub mod parse_accounts;
 pub mod parse_address_lookup_table;
 pub mod parse_associated_token;
@@ -53,12 +64,6 @@ pub struct BlockEncodingOptions {
     pub transaction_details: TransactionDetails,
     pub show_rewards: bool,
     pub max_supported_transaction_version: Option<u8>,
-}
-
-#[derive(Error, Debug, PartialEq, Eq, Clone)]
-pub enum EncodeError {
-    #[error("Encoding does not support transaction version {0}")]
-    UnsupportedTransactionVersion(u8),
 }
 
 /// Represents types that can be encoded into one of several encoding formats
@@ -83,152 +88,36 @@ trait JsonAccounts {
     fn build_json_accounts(&self) -> Self::Encoded;
 }
 
-#[derive(Serialize, Deserialize, Clone, Copy, Debug, Eq, Hash, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub enum TransactionBinaryEncoding {
-    Base58,
-    Base64,
-}
-
-#[derive(Serialize, Deserialize, Clone, Copy, Debug, Eq, Hash, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub enum UiTransactionEncoding {
-    Binary, // Legacy. Retained for RPC backwards compatibility
-    Base64,
-    Base58,
-    Json,
-    JsonParsed,
-}
-
-impl UiTransactionEncoding {
-    pub fn into_binary_encoding(&self) -> Option<TransactionBinaryEncoding> {
-        match self {
-            Self::Binary | Self::Base58 => Some(TransactionBinaryEncoding::Base58),
-            Self::Base64 => Some(TransactionBinaryEncoding::Base64),
-            _ => None,
-        }
+fn make_ui_partially_decoded_instruction(
+    instruction: &CompiledInstruction,
+    account_keys: &AccountKeys,
+    stack_height: Option<u32>,
+) -> UiPartiallyDecodedInstruction {
+    UiPartiallyDecodedInstruction {
+        program_id: account_keys[instruction.program_id_index as usize].to_string(),
+        accounts: instruction
+            .accounts
+            .iter()
+            .map(|&i| account_keys[i as usize].to_string())
+            .collect(),
+        data: bs58::encode(instruction.data.clone()).into_string(),
+        stack_height,
     }
 }
 
-impl fmt::Display for UiTransactionEncoding {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let v = serde_json::to_value(self).map_err(|_| fmt::Error)?;
-        let s = v.as_str().ok_or(fmt::Error)?;
-        write!(f, "{s}")
+pub fn parse_ui_instruction(
+    instruction: &CompiledInstruction,
+    account_keys: &AccountKeys,
+    stack_height: Option<u32>,
+) -> UiInstruction {
+    let program_id = &account_keys[instruction.program_id_index as usize];
+    if let Ok(parsed_instruction) = parse(program_id, instruction, account_keys, stack_height) {
+        UiInstruction::Parsed(UiParsedInstruction::Parsed(parsed_instruction))
+    } else {
+        UiInstruction::Parsed(UiParsedInstruction::PartiallyDecoded(
+            make_ui_partially_decoded_instruction(instruction, account_keys, stack_height),
+        ))
     }
-}
-
-#[derive(Debug, Clone, Copy, Eq, Hash, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub enum TransactionDetails {
-    Full,
-    Signatures,
-    None,
-    Accounts,
-}
-
-impl Default for TransactionDetails {
-    fn default() -> Self {
-        Self::Full
-    }
-}
-
-/// A duplicate representation of an Instruction for pretty JSON serialization
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase", untagged)]
-pub enum UiInstruction {
-    Compiled(UiCompiledInstruction),
-    Parsed(UiParsedInstruction),
-}
-
-impl UiInstruction {
-    pub fn parse(
-        instruction: &CompiledInstruction,
-        account_keys: &AccountKeys,
-        stack_height: Option<u32>,
-    ) -> Self {
-        let program_id = &account_keys[instruction.program_id_index as usize];
-        if let Ok(parsed_instruction) = parse(program_id, instruction, account_keys, stack_height) {
-            UiInstruction::Parsed(UiParsedInstruction::Parsed(parsed_instruction))
-        } else {
-            UiInstruction::Parsed(UiParsedInstruction::PartiallyDecoded(
-                UiPartiallyDecodedInstruction::from(instruction, account_keys, stack_height),
-            ))
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase", untagged)]
-pub enum UiParsedInstruction {
-    Parsed(ParsedInstruction),
-    PartiallyDecoded(UiPartiallyDecodedInstruction),
-}
-
-/// A duplicate representation of a CompiledInstruction for pretty JSON serialization
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UiCompiledInstruction {
-    pub program_id_index: u8,
-    pub accounts: Vec<u8>,
-    pub data: String,
-    pub stack_height: Option<u32>,
-}
-
-impl UiCompiledInstruction {
-    fn from(instruction: &CompiledInstruction, stack_height: Option<u32>) -> Self {
-        Self {
-            program_id_index: instruction.program_id_index,
-            accounts: instruction.accounts.clone(),
-            data: bs58::encode(&instruction.data).into_string(),
-            stack_height,
-        }
-    }
-}
-
-/// A partially decoded CompiledInstruction that includes explicit account addresses
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UiPartiallyDecodedInstruction {
-    pub program_id: String,
-    pub accounts: Vec<String>,
-    pub data: String,
-    pub stack_height: Option<u32>,
-}
-
-impl UiPartiallyDecodedInstruction {
-    fn from(
-        instruction: &CompiledInstruction,
-        account_keys: &AccountKeys,
-        stack_height: Option<u32>,
-    ) -> Self {
-        Self {
-            program_id: account_keys[instruction.program_id_index as usize].to_string(),
-            accounts: instruction
-                .accounts
-                .iter()
-                .map(|&i| account_keys[i as usize].to_string())
-                .collect(),
-            data: bs58::encode(instruction.data.clone()).into_string(),
-            stack_height,
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct InnerInstructions {
-    /// Transaction instruction index
-    pub index: u8,
-    /// List of inner instructions
-    pub instructions: Vec<InnerInstruction>,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct InnerInstruction {
-    /// Compiled instruction
-    pub instruction: CompiledInstruction,
-    /// Invocation stack height of the instruction,
-    pub stack_height: Option<u32>,
 }
 
 /// Maps a list of inner instructions from `solana_sdk` into a list of this
@@ -252,314 +141,92 @@ pub fn map_inner_instructions(
         .filter(|i| !i.instructions.is_empty())
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UiInnerInstructions {
-    /// Transaction instruction index
-    pub index: u8,
-    /// List of inner instructions
-    pub instructions: Vec<UiInstruction>,
-}
-
-impl UiInnerInstructions {
-    pub fn parse(inner_instructions: InnerInstructions, account_keys: &AccountKeys) -> Self {
-        Self {
-            index: inner_instructions.index,
-            instructions: inner_instructions
-                .instructions
-                .iter()
-                .map(
-                    |InnerInstruction {
-                         instruction: ix,
-                         stack_height,
-                     }| {
-                        UiInstruction::parse(ix, account_keys, *stack_height)
-                    },
-                )
-                .collect(),
-        }
+pub fn parse_ui_inner_instructions(
+    inner_instructions: InnerInstructions,
+    account_keys: &AccountKeys,
+) -> UiInnerInstructions {
+    UiInnerInstructions {
+        index: inner_instructions.index,
+        instructions: inner_instructions
+            .instructions
+            .iter()
+            .map(
+                |InnerInstruction {
+                     instruction: ix,
+                     stack_height,
+                 }| { parse_ui_instruction(ix, account_keys, *stack_height) },
+            )
+            .collect(),
     }
 }
 
-impl From<InnerInstructions> for UiInnerInstructions {
-    fn from(inner_instructions: InnerInstructions) -> Self {
-        Self {
-            index: inner_instructions.index,
-            instructions: inner_instructions
-                .instructions
-                .iter()
-                .map(
-                    |InnerInstruction {
-                         instruction: ix,
-                         stack_height,
-                     }| {
-                        UiInstruction::Compiled(UiCompiledInstruction::from(ix, *stack_height))
-                    },
-                )
-                .collect(),
-        }
+fn build_simple_ui_transaction_status_meta(
+    meta: TransactionStatusMeta,
+    show_rewards: bool,
+) -> UiTransactionStatusMeta {
+    UiTransactionStatusMeta {
+        err: meta.status.clone().err(),
+        status: meta.status,
+        fee: meta.fee,
+        pre_balances: meta.pre_balances,
+        post_balances: meta.post_balances,
+        inner_instructions: OptionSerializer::Skip,
+        log_messages: OptionSerializer::Skip,
+        pre_token_balances: meta
+            .pre_token_balances
+            .map(|balance| balance.into_iter().map(Into::into).collect())
+            .into(),
+        post_token_balances: meta
+            .post_token_balances
+            .map(|balance| balance.into_iter().map(Into::into).collect())
+            .into(),
+        rewards: if show_rewards {
+            meta.rewards.into()
+        } else {
+            OptionSerializer::Skip
+        },
+        loaded_addresses: OptionSerializer::Skip,
+        return_data: OptionSerializer::Skip,
+        compute_units_consumed: OptionSerializer::Skip,
     }
 }
 
-#[derive(Clone, Debug, PartialEq)]
-pub struct TransactionTokenBalance {
-    pub account_index: u8,
-    pub mint: String,
-    pub ui_token_amount: UiTokenAmount,
-    pub owner: String,
-    pub program_id: String,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UiTransactionTokenBalance {
-    pub account_index: u8,
-    pub mint: String,
-    pub ui_token_amount: UiTokenAmount,
-    #[serde(
-        default = "OptionSerializer::skip",
-        skip_serializing_if = "OptionSerializer::should_skip"
-    )]
-    pub owner: OptionSerializer<String>,
-    #[serde(
-        default = "OptionSerializer::skip",
-        skip_serializing_if = "OptionSerializer::should_skip"
-    )]
-    pub program_id: OptionSerializer<String>,
-}
-
-impl From<TransactionTokenBalance> for UiTransactionTokenBalance {
-    fn from(token_balance: TransactionTokenBalance) -> Self {
-        Self {
-            account_index: token_balance.account_index,
-            mint: token_balance.mint,
-            ui_token_amount: token_balance.ui_token_amount,
-            owner: if !token_balance.owner.is_empty() {
-                OptionSerializer::Some(token_balance.owner)
-            } else {
-                OptionSerializer::Skip
-            },
-            program_id: if !token_balance.program_id.is_empty() {
-                OptionSerializer::Some(token_balance.program_id)
-            } else {
-                OptionSerializer::Skip
-            },
-        }
+fn parse_ui_transaction_status_meta(
+    meta: TransactionStatusMeta,
+    static_keys: &[Pubkey],
+    show_rewards: bool,
+) -> UiTransactionStatusMeta {
+    let account_keys = AccountKeys::new(static_keys, Some(&meta.loaded_addresses));
+    UiTransactionStatusMeta {
+        err: meta.status.clone().err(),
+        status: meta.status,
+        fee: meta.fee,
+        pre_balances: meta.pre_balances,
+        post_balances: meta.post_balances,
+        inner_instructions: meta
+            .inner_instructions
+            .map(|ixs| {
+                ixs.into_iter()
+                    .map(|ix| parse_ui_inner_instructions(ix, &account_keys))
+                    .collect()
+            })
+            .into(),
+        log_messages: meta.log_messages.into(),
+        pre_token_balances: meta
+            .pre_token_balances
+            .map(|balance| balance.into_iter().map(Into::into).collect())
+            .into(),
+        post_token_balances: meta
+            .post_token_balances
+            .map(|balance| balance.into_iter().map(Into::into).collect())
+            .into(),
+        rewards: if show_rewards { meta.rewards } else { None }.into(),
+        loaded_addresses: OptionSerializer::Skip,
+        return_data: OptionSerializer::or_skip(
+            meta.return_data.map(|return_data| return_data.into()),
+        ),
+        compute_units_consumed: OptionSerializer::or_skip(meta.compute_units_consumed),
     }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct TransactionStatusMeta {
-    pub status: TransactionResult<()>,
-    pub fee: u64,
-    pub pre_balances: Vec<u64>,
-    pub post_balances: Vec<u64>,
-    pub inner_instructions: Option<Vec<InnerInstructions>>,
-    pub log_messages: Option<Vec<String>>,
-    pub pre_token_balances: Option<Vec<TransactionTokenBalance>>,
-    pub post_token_balances: Option<Vec<TransactionTokenBalance>>,
-    pub rewards: Option<Rewards>,
-    pub loaded_addresses: LoadedAddresses,
-    pub return_data: Option<TransactionReturnData>,
-    pub compute_units_consumed: Option<u64>,
-}
-
-impl Default for TransactionStatusMeta {
-    fn default() -> Self {
-        Self {
-            status: Ok(()),
-            fee: 0,
-            pre_balances: vec![],
-            post_balances: vec![],
-            inner_instructions: None,
-            log_messages: None,
-            pre_token_balances: None,
-            post_token_balances: None,
-            rewards: None,
-            loaded_addresses: LoadedAddresses::default(),
-            return_data: None,
-            compute_units_consumed: None,
-        }
-    }
-}
-
-/// A duplicate representation of TransactionStatusMeta with `err` field
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UiTransactionStatusMeta {
-    pub err: Option<TransactionError>,
-    pub status: TransactionResult<()>, // This field is deprecated.  See https://github.com/solana-labs/solana/issues/9302
-    pub fee: u64,
-    pub pre_balances: Vec<u64>,
-    pub post_balances: Vec<u64>,
-    #[serde(
-        default = "OptionSerializer::none",
-        skip_serializing_if = "OptionSerializer::should_skip"
-    )]
-    pub inner_instructions: OptionSerializer<Vec<UiInnerInstructions>>,
-    #[serde(
-        default = "OptionSerializer::none",
-        skip_serializing_if = "OptionSerializer::should_skip"
-    )]
-    pub log_messages: OptionSerializer<Vec<String>>,
-    #[serde(
-        default = "OptionSerializer::none",
-        skip_serializing_if = "OptionSerializer::should_skip"
-    )]
-    pub pre_token_balances: OptionSerializer<Vec<UiTransactionTokenBalance>>,
-    #[serde(
-        default = "OptionSerializer::none",
-        skip_serializing_if = "OptionSerializer::should_skip"
-    )]
-    pub post_token_balances: OptionSerializer<Vec<UiTransactionTokenBalance>>,
-    #[serde(
-        default = "OptionSerializer::none",
-        skip_serializing_if = "OptionSerializer::should_skip"
-    )]
-    pub rewards: OptionSerializer<Rewards>,
-    #[serde(
-        default = "OptionSerializer::skip",
-        skip_serializing_if = "OptionSerializer::should_skip"
-    )]
-    pub loaded_addresses: OptionSerializer<UiLoadedAddresses>,
-    #[serde(
-        default = "OptionSerializer::skip",
-        skip_serializing_if = "OptionSerializer::should_skip"
-    )]
-    pub return_data: OptionSerializer<UiTransactionReturnData>,
-    #[serde(
-        default = "OptionSerializer::skip",
-        skip_serializing_if = "OptionSerializer::should_skip"
-    )]
-    pub compute_units_consumed: OptionSerializer<u64>,
-}
-
-/// A duplicate representation of LoadedAddresses
-#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UiLoadedAddresses {
-    pub writable: Vec<String>,
-    pub readonly: Vec<String>,
-}
-
-impl From<&LoadedAddresses> for UiLoadedAddresses {
-    fn from(loaded_addresses: &LoadedAddresses) -> Self {
-        Self {
-            writable: loaded_addresses
-                .writable
-                .iter()
-                .map(ToString::to_string)
-                .collect(),
-            readonly: loaded_addresses
-                .readonly
-                .iter()
-                .map(ToString::to_string)
-                .collect(),
-        }
-    }
-}
-
-impl UiTransactionStatusMeta {
-    fn parse(meta: TransactionStatusMeta, static_keys: &[Pubkey], show_rewards: bool) -> Self {
-        let account_keys = AccountKeys::new(static_keys, Some(&meta.loaded_addresses));
-        Self {
-            err: meta.status.clone().err(),
-            status: meta.status,
-            fee: meta.fee,
-            pre_balances: meta.pre_balances,
-            post_balances: meta.post_balances,
-            inner_instructions: meta
-                .inner_instructions
-                .map(|ixs| {
-                    ixs.into_iter()
-                        .map(|ix| UiInnerInstructions::parse(ix, &account_keys))
-                        .collect()
-                })
-                .into(),
-            log_messages: meta.log_messages.into(),
-            pre_token_balances: meta
-                .pre_token_balances
-                .map(|balance| balance.into_iter().map(Into::into).collect())
-                .into(),
-            post_token_balances: meta
-                .post_token_balances
-                .map(|balance| balance.into_iter().map(Into::into).collect())
-                .into(),
-            rewards: if show_rewards { meta.rewards } else { None }.into(),
-            loaded_addresses: OptionSerializer::Skip,
-            return_data: OptionSerializer::or_skip(
-                meta.return_data.map(|return_data| return_data.into()),
-            ),
-            compute_units_consumed: OptionSerializer::or_skip(meta.compute_units_consumed),
-        }
-    }
-
-    fn build_simple(meta: TransactionStatusMeta, show_rewards: bool) -> Self {
-        Self {
-            err: meta.status.clone().err(),
-            status: meta.status,
-            fee: meta.fee,
-            pre_balances: meta.pre_balances,
-            post_balances: meta.post_balances,
-            inner_instructions: OptionSerializer::Skip,
-            log_messages: OptionSerializer::Skip,
-            pre_token_balances: meta
-                .pre_token_balances
-                .map(|balance| balance.into_iter().map(Into::into).collect())
-                .into(),
-            post_token_balances: meta
-                .post_token_balances
-                .map(|balance| balance.into_iter().map(Into::into).collect())
-                .into(),
-            rewards: if show_rewards {
-                meta.rewards.into()
-            } else {
-                OptionSerializer::Skip
-            },
-            loaded_addresses: OptionSerializer::Skip,
-            return_data: OptionSerializer::Skip,
-            compute_units_consumed: OptionSerializer::Skip,
-        }
-    }
-}
-
-impl From<TransactionStatusMeta> for UiTransactionStatusMeta {
-    fn from(meta: TransactionStatusMeta) -> Self {
-        Self {
-            err: meta.status.clone().err(),
-            status: meta.status,
-            fee: meta.fee,
-            pre_balances: meta.pre_balances,
-            post_balances: meta.post_balances,
-            inner_instructions: meta
-                .inner_instructions
-                .map(|ixs| ixs.into_iter().map(Into::into).collect())
-                .into(),
-            log_messages: meta.log_messages.into(),
-            pre_token_balances: meta
-                .pre_token_balances
-                .map(|balance| balance.into_iter().map(Into::into).collect())
-                .into(),
-            post_token_balances: meta
-                .post_token_balances
-                .map(|balance| balance.into_iter().map(Into::into).collect())
-                .into(),
-            rewards: meta.rewards.into(),
-            loaded_addresses: Some(UiLoadedAddresses::from(&meta.loaded_addresses)).into(),
-            return_data: OptionSerializer::or_skip(
-                meta.return_data.map(|return_data| return_data.into()),
-            ),
-            compute_units_consumed: OptionSerializer::or_skip(meta.compute_units_consumed),
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub enum TransactionConfirmationStatus {
-    Processed,
-    Confirmed,
-    Finalized,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
@@ -606,27 +273,6 @@ impl TransactionStatus {
         }
     }
 }
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct ConfirmedTransactionStatusWithSignature {
-    pub signature: Signature,
-    pub slot: Slot,
-    pub err: Option<TransactionError>,
-    pub memo: Option<String>,
-    pub block_time: Option<UnixTimestamp>,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Reward {
-    pub pubkey: String,
-    pub lamports: i64,
-    pub post_balance: u64, // Account balance in lamports after `lamports` was applied
-    pub reward_type: Option<RewardType>,
-    pub commission: Option<u8>, // Vote account commission when the reward was credited, only present for voting and staking rewards
-}
-
-pub type Rewards = Vec<Reward>;
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct RewardsAndNumPartitions {
@@ -813,24 +459,6 @@ impl From<UiConfirmedBlock> for EncodedConfirmedBlock {
     }
 }
 
-#[derive(Debug, PartialEq, Serialize, Deserialize, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct UiConfirmedBlock {
-    pub previous_blockhash: String,
-    pub blockhash: String,
-    pub parent_slot: Slot,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub transactions: Option<Vec<EncodedTransactionWithStatusMeta>>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub signatures: Option<Vec<String>>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub rewards: Option<Rewards>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub num_reward_partitions: Option<u64>,
-    pub block_time: Option<UnixTimestamp>,
-    pub block_height: Option<u64>,
-}
-
 // Confirmed block with type guarantees that transaction metadata is always
 // present, as well as a list of the entry data needed to cryptographically
 // verify the block. Used for uploading to BigTable.
@@ -966,7 +594,7 @@ impl VersionedTransactionWithStatusMeta {
         Ok(EncodedTransactionWithStatusMeta {
             transaction: self.transaction.encode_with_meta(encoding, &self.meta),
             meta: Some(match encoding {
-                UiTransactionEncoding::JsonParsed => UiTransactionStatusMeta::parse(
+                UiTransactionEncoding::JsonParsed => parse_ui_transaction_status_meta(
                     self.meta,
                     self.transaction.message.static_account_keys(),
                     show_rewards,
@@ -1020,22 +648,13 @@ impl VersionedTransactionWithStatusMeta {
                     .collect(),
                 account_keys,
             }),
-            meta: Some(UiTransactionStatusMeta::build_simple(
+            meta: Some(build_simple_ui_transaction_status_meta(
                 self.meta,
                 show_rewards,
             )),
             version,
         })
     }
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct EncodedTransactionWithStatusMeta {
-    pub transaction: EncodedTransaction,
-    pub meta: Option<UiTransactionStatusMeta>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub version: Option<TransactionVersion>,
 }
 
 #[derive(Debug, Clone, PartialEq)]
@@ -1081,15 +700,6 @@ pub struct EncodedConfirmedTransactionWithStatusMeta {
     #[serde(flatten)]
     pub transaction: EncodedTransactionWithStatusMeta,
     pub block_time: Option<UnixTimestamp>,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase", untagged)]
-pub enum EncodedTransaction {
-    LegacyBinary(String), // Old way of expressing base-58, retained for RPC backwards compatibility
-    Binary(String, TransactionBinaryEncoding),
-    Json(UiTransaction),
-    Accounts(UiAccountsList),
 }
 
 impl EncodableWithMeta for VersionedTransaction {
@@ -1203,44 +813,6 @@ impl JsonAccounts for Transaction {
     }
 }
 
-impl EncodedTransaction {
-    pub fn decode(&self) -> Option<VersionedTransaction> {
-        let (blob, encoding) = match self {
-            Self::Json(_) | Self::Accounts(_) => return None,
-            Self::LegacyBinary(blob) => (blob, TransactionBinaryEncoding::Base58),
-            Self::Binary(blob, encoding) => (blob, *encoding),
-        };
-
-        let transaction: Option<VersionedTransaction> = match encoding {
-            TransactionBinaryEncoding::Base58 => bs58::decode(blob)
-                .into_vec()
-                .ok()
-                .and_then(|bytes| bincode::deserialize(&bytes).ok()),
-            TransactionBinaryEncoding::Base64 => BASE64_STANDARD
-                .decode(blob)
-                .ok()
-                .and_then(|bytes| bincode::deserialize(&bytes).ok()),
-        };
-
-        transaction.filter(|transaction| transaction.sanitize().is_ok())
-    }
-}
-
-/// A duplicate representation of a Transaction for pretty JSON serialization
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UiTransaction {
-    pub signatures: Vec<String>,
-    pub message: UiMessage,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase", untagged)]
-pub enum UiMessage {
-    Parsed(UiParsedMessage),
-    Raw(UiRawMessage),
-}
-
 impl Encodable for Message {
     type Encoded = UiMessage;
     fn encode(&self, encoding: UiTransactionEncoding) -> Self::Encoded {
@@ -1252,7 +824,7 @@ impl Encodable for Message {
                 instructions: self
                     .instructions
                     .iter()
-                    .map(|instruction| UiInstruction::parse(instruction, &account_keys, None))
+                    .map(|instruction| parse_ui_instruction(instruction, &account_keys, None))
                     .collect(),
                 address_table_lookups: None,
             })
@@ -1286,7 +858,7 @@ impl Encodable for v0::Message {
                 instructions: self
                     .instructions
                     .iter()
-                    .map(|instruction| UiInstruction::parse(instruction, &account_keys, None))
+                    .map(|instruction| parse_ui_instruction(instruction, &account_keys, None))
                     .collect(),
                 address_table_lookups: None,
             })
@@ -1327,7 +899,7 @@ impl EncodableWithMeta for v0::Message {
                 instructions: self
                     .instructions
                     .iter()
-                    .map(|instruction| UiInstruction::parse(instruction, &account_keys, None))
+                    .map(|instruction| parse_ui_instruction(instruction, &account_keys, None))
                     .collect(),
                 address_table_lookups: Some(
                     self.address_table_lookups.iter().map(Into::into).collect(),
@@ -1354,55 +926,6 @@ impl EncodableWithMeta for v0::Message {
     }
 }
 
-/// A duplicate representation of a Message, in raw format, for pretty JSON serialization
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UiRawMessage {
-    pub header: MessageHeader,
-    pub account_keys: Vec<String>,
-    pub recent_blockhash: String,
-    pub instructions: Vec<UiCompiledInstruction>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub address_table_lookups: Option<Vec<UiAddressTableLookup>>,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UiAccountsList {
-    pub signatures: Vec<String>,
-    pub account_keys: Vec<ParsedAccount>,
-}
-
-/// A duplicate representation of a MessageAddressTableLookup, in raw format, for pretty JSON serialization
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UiAddressTableLookup {
-    pub account_key: String,
-    pub writable_indexes: Vec<u8>,
-    pub readonly_indexes: Vec<u8>,
-}
-
-impl From<&MessageAddressTableLookup> for UiAddressTableLookup {
-    fn from(lookup: &MessageAddressTableLookup) -> Self {
-        Self {
-            account_key: lookup.account_key.to_string(),
-            writable_indexes: lookup.writable_indexes.clone(),
-            readonly_indexes: lookup.readonly_indexes.clone(),
-        }
-    }
-}
-
-/// A duplicate representation of a Message, in parsed format, for pretty JSON serialization
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UiParsedMessage {
-    pub account_keys: Vec<ParsedAccount>,
-    pub recent_blockhash: String,
-    pub instructions: Vec<UiInstruction>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub address_table_lookups: Option<Vec<UiAddressTableLookup>>,
-}
-
 // A serialized `Vec<TransactionByAddrInfo>` is stored in the `tx-by-addr` table.  The row keys are
 // the one's compliment of the slot so that rows may be listed in reverse order
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
@@ -1412,40 +935,6 @@ pub struct TransactionByAddrInfo {
     pub index: u32,                    // Where the transaction is located in the block
     pub memo: Option<String>,          // Transaction memo
     pub block_time: Option<UnixTimestamp>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub struct UiTransactionReturnData {
-    pub program_id: String,
-    pub data: (String, UiReturnDataEncoding),
-}
-
-impl Default for UiTransactionReturnData {
-    fn default() -> Self {
-        Self {
-            program_id: String::default(),
-            data: (String::default(), UiReturnDataEncoding::Base64),
-        }
-    }
-}
-
-impl From<TransactionReturnData> for UiTransactionReturnData {
-    fn from(return_data: TransactionReturnData) -> Self {
-        Self {
-            program_id: return_data.program_id.to_string(),
-            data: (
-                BASE64_STANDARD.encode(return_data.data),
-                UiReturnDataEncoding::Base64,
-            ),
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Clone, Copy, Debug, Eq, Hash, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub enum UiReturnDataEncoding {
-    Base64,
 }
 
 #[cfg(test)]
@@ -1662,13 +1151,13 @@ mod test {
         }",
         )
         .unwrap();
-        let ui_meta_parse_with_rewards = UiTransactionStatusMeta::parse(meta.clone(), &[], true);
+        let ui_meta_parse_with_rewards = parse_ui_transaction_status_meta(meta.clone(), &[], true);
         assert_eq!(
             serde_json::to_value(ui_meta_parse_with_rewards).unwrap(),
             expected_json_output_value
         );
 
-        let ui_meta_parse_no_rewards = UiTransactionStatusMeta::parse(meta, &[], false);
+        let ui_meta_parse_no_rewards = parse_ui_transaction_status_meta(meta, &[], false);
         assert_eq!(
             serde_json::to_value(ui_meta_parse_no_rewards).unwrap(),
             expected_json_output_value
